refactor: change package for SocketAppHandler
