+++ zio-http/src/main/scala/zhttp/service/server/WebSocketUpgrade.scala
@@ -4,7 +4,7 @@ import io.netty.channel.{ChannelHandler, ChannelHandlerContext}
 import io.netty.handler.codec.http._
 import io.netty.handler.codec.http.websocketx.WebSocketServerProtocolHandler
 import zhttp.http.{Response, Status}
-import zhttp.service.{HttpRuntime, WEB_SOCKET_HANDLER}
+import zhttp.service.{HttpRuntime, SocketAppHandler, WEB_SOCKET_HANDLER}
 
 /**
  * Module to switch protocol to websockets
